Fix de tipos de datos

+++ src/main/java/tech/telefonicachile/kpibackendapi/dtos/response/PrtTotalIncReqCumplidosResponse.java
@@ -4,15 +4,17 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @Data
 @Builder
 public class PrtTotalIncReqCumplidosResponse {
     Integer idGrupoAsignacion;
     String grupoAsignacion;
-    Long totalCumplidos;
+    BigDecimal totalCumplidos;
 
-    public PrtTotalIncReqCumplidosResponse(Integer idGrupoAsignacion, String grupoAsignacion, Long totalCumplidos) {
+    public PrtTotalIncReqCumplidosResponse(Integer idGrupoAsignacion, String grupoAsignacion, BigDecimal totalCumplidos) {
         this.idGrupoAsignacion = idGrupoAsignacion;
         this.grupoAsignacion = grupoAsignacion;
         this.totalCumplidos = totalCumplidos;
